pyproject.toml
[project]
name = "resgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.11.0",
    "google-genai>=1.18.0",
    "pymupdf4llm>=0.0.24",
    "requests>=2.32.3",
    "trafilatura>=2.0.0",
]
